Moves dao and schema to different files, changes vote to short
